requirements.txt
psutil==5.9.8
GPUtil==1.4.0
influxdb-client==1.38.0
flask==3.0.0
pyyaml==6.0.1
aiohttp==3.9.1
jinja2==3.1.2
python-json-logger==2.0.7
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
